refactored, and added same single browser-per-thread for pdf
